requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
distro==1.9.0
environs==14.1.0
fastapi==0.115.6
fastapi_cors==0.0.6
groq==0.14.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
marshmallow==3.25.0
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4